Backend/requirements.txt
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
blinker==1.8.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7.post1
decorator==5.1.1
Deprecated==1.2.14
docopt==0.6.2
executing==1.2.0
fastjsonschema==2.18.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Limiter==3.7.0
fonttools==4.42.1
idna==3.4
importlib_resources==6.4.0
iniconfig==2.0.0
ipykernel==6.25.1
ipython==8.15.0
itsdangerous==2.2.0
jedi==0.19.0
Jinja2==3.1.4
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client==8.3.1
jupyter_core==5.3.1
kiwisolver==1.4.5
limits==3.12.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mplcursors==0.5.2
musicbrainzngs==0.7.1
nbformat==5.9.2
nest-asyncio==1.5.7
numpy==1.25.2
ordered-set==4.1.0
packaging==23.1
pandas==2.1.0
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
pipreqs==0.4.13
platformdirs==3.10.0
plotly==5.16.1
pluggy==1.5.0
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.16.1
pyparsing==3.0.9
pytest==8.2.2
pytest-mock==3.14.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
rich==13.7.1
rpds-py==0.10.0
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.12.2
six==1.16.0
stack-data==0.6.2
tenacity==8.2.3
threadpoolctl==3.2.0
tornado==6.3.3
traitlets==5.9.0
typing_extensions==4.11.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
wcwidth==0.2.6
Werkzeug==3.0.3
wrapt==1.16.0
yarg==0.1.9
